+++ bin/.gitignore
@@ -1,2 +1,2 @@
+/App/
 /Ex3/
-/Task/
