--- bin/App/.gitignore
@@ -1,2 +1,0 @@
-/mainFraim.class
-/package-info.class
